# Only generate boundaries within settlements.