--- a/backend/benches/boundary_stats.rs
+++ b/backend/benches/boundary_stats.rs
@@ -4,8 +4,8 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 
 fn benchmark_build_map_model(c: &mut Criterion) {
     for (fixture, expected_population_zones, expected_generated_boundaries) in [
-        (NeighbourhoodFixture::INVERNESS, 325, 8958),
-        (NeighbourhoodFixture::DUNDEE, 200, 240),
+        (NeighbourhoodFixture::INVERNESS, 325, 390),
+        (NeighbourhoodFixture::DUNDEE, 200, 232),
     ] {
         // Do the file i/o (reading OSM.xml) outside of the bench loop
         let (neighbourhood, map) = fixture.neighbourhood_map().unwrap();