An attempt at https://github.com/a-b-street/ltn/issues/254#issuecomment-2744699163

FIXES #158 
FIXES #254

**before:** Largely rural areas (Highlands) had many huge, improbably, suggested boundaries, and ~30s load time.

<img width="785" alt="Screenshot 2025-03-25 at 09 18 28" src="https://github.com/user-attachments/assets/c3790270-db02-450e-9613-885dc69f73c9" />

Though it did look kind of fun.

**after:**
<img width="808" alt="Screenshot 2025-03-25 at 09 18 08" src="https://github.com/user-attachments/assets/9746f4e8-c35e-4744-bb01-f059f0456bc5" />

Zooming in to Inverness
<img width="1388" alt="Screenshot 2025-03-25 at 09 18 15" src="https://github.com/user-attachments/assets/6db9ee28-a04b-4bb4-a80d-03aeeb99a16c" />


Denser areas, e.g. Glasgow City are more or less unaffected by this change:

**before:**
<img width="758" alt="Screenshot 2025-03-25 at 09 19 42" src="https://github.com/user-attachments/assets/86252a7c-384c-42a5-a0d5-ebd38bb75f55" />

**after:** Some of the outskirt neighborhoods seem to have an additional split, presumably when a previously "severance based" boundary is now split by the settlement boundary. I think that's fine:

<img width="1091" alt="Screenshot 2025-03-25 at 09 19 25" src="https://github.com/user-attachments/assets/1cafdf62-a190-4b3c-9116-c6fb1ff59e38" />

If we decide to, I think adding buffering would be a straight forward change on top of this. But in my own QA, I haven't yet found cases where it would be helpful.

Performance diff:

```
inverness/generate auto boundaries (and stats)
                        time:   [4.5871 s 4.6247 s 4.6708 s]
                        change: [-79.509% -78.969% -78.414%] (p = 0.00 < 0.05)
                        Performance has improved.
dundee/generate auto boundaries (and stats)
                        time:   [147.92 ms 148.44 ms 149.01 ms]
                        change: [+107.50% +108.33% +109.26%] (p = 0.00 < 0.05)
                        Performance has regressed.
```

The fast ones got quite a bit slower (relatively), but I think are still fast enough. While the slow ones got a **lot** faster. I think this puts us into an acceptable range for Highlands.