@michaelkirk 2025-03-22
One tricky bit are the settlement outskirts. 

So we split the LAD into boundary polygons using severances as the splitters. However many areas on the inner edge of the settlements ("outskirts"), end up in the huge mostly "unsettled" severance polygon. That one, being huge and representing the negative space (mostly) between settlements, gets removed. Meaning, the outskirts are removed as collateral damage.

My attempt at addressing this was to introduce the settlement boundaries themselves as a splitter, along with the traditional "severances".

**splitting by severances only:**

<img width="571" alt="arboroath-no-outskirts" src="https://github.com/user-attachments/assets/557cf762-cfc3-4b4e-8284-3574537598d9" />

**splitting by severances and settlement boundary:**
<img width="517" alt="arboroath-with-outskirts" src="https://github.com/user-attachments/assets/723c3c6f-a6e5-4ef0-99fc-0521dc369b1d" />