@michaelkirk 2025-03-25
I did a little cleanup and updated the description. I think this is a big improvement. I think we should merge it and get sustrans to try it out.